collect_traffic_stack/collect_traffic_lambda/requirements.txt
githubpy==1.1.0
simplejson==3.17.0
dynamodb_encryption_sdk
